# feat: use interpolated string in `RightHandSide`